TPI/robot_narrativo/requirements.txt
flask>=2.3.0
gtts>=2.3.0
pyttsx3>=2.90
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0

